+++ examples/example.php
@@ -2,6 +2,7 @@
 
 use Aza\Components\CliBase\Base;
 use Aza\Components\Thread\Exceptions\Exception;
+use Aza\Components\Thread\SimpleThread;
 use Aza\Components\Thread\Thread;
 use Aza\Components\Thread\ThreadPool;
 
@@ -26,9 +27,7 @@ require __DIR__ . '/../vendor/autoload.php';
 class TestThreadReturnFirstArgument extends Thread
 {
 	/**
-	 * Main processing.
-	 *
-	 * @return mixed
+	 * {@inheritdoc}
 	 */
 	function process()
 	{
@@ -44,9 +43,7 @@ class TestThreadEvents extends Thread
 	const EV_PROCESS = 'process';
 
 	/**
-	 * Main processing.
-	 *
-	 * @return mixed
+	 * {@inheritdoc}
 	 */
 	function process()
 	{
@@ -73,8 +70,8 @@ if (!Thread::$useForks) {
 
 // ----------------------------------------------
 echo PHP_EOL,
-     'Simple example with one thread',
-     PHP_EOL;
+	'Simple example with one thread',
+	PHP_EOL;
 
 $num = 10; // Number of tasks
 $thread = new TestThreadReturnFirstArgument();
@@ -103,8 +100,8 @@ $thread->cleanup();
 
 // ----------------------------------------------
 echo PHP_EOL,
-     'Simple example with thread events',
-     PHP_EOL;
+	'Simple example with thread events',
+	PHP_EOL;
 
 $events = 10; // Number of events
 $num    = 3;  // Number of tasks
@@ -132,8 +129,8 @@ $thread->cleanup();
 $threads = 4;
 
 echo PHP_EOL,
-     "Simple example with pool of threads ($threads)",
-     PHP_EOL;
+	"Simple example with pool of threads ($threads)",
+	PHP_EOL;
 
 $pool = new ThreadPool('TestThreadReturnFirstArgument', $threads);
 
@@ -173,8 +170,8 @@ $jobs     = range(1, 30);
 $jobs_num = count($jobs);
 
 echo PHP_EOL,
-     "Example with pool of threads ($threads) and pool of jobs ($jobs_num)",
-     PHP_EOL;
+	"Example with pool of threads ($threads) and pool of jobs ($jobs_num)",
+	PHP_EOL;
 
 $pool = new ThreadPool('TestThreadReturnFirstArgument', $threads);
 
@@ -212,3 +209,35 @@ do {
 	}
 } while ($num > 0);
 $pool->cleanup();
+
+
+
+// ----------------------------------------------
+echo PHP_EOL,
+	'Simple example with one "closure" thread',
+	PHP_EOL;
+
+$num = 10; // Number of tasks
+$thread = SimpleThread::create(function($arg) {
+	return $arg;
+});
+
+// "closure" threads are not preforked by default
+// and not multitask too. You can change it via
+// the second argument of `SimpleThread::create`.
+
+for ($i = 0; $i < $num; $i++) {
+	$value = $i;
+	// Run task and wait for the result
+	if ($thread->run($value)->wait()->getSuccess()) {
+		// Success
+		$result = $thread->getResult();
+		echo 'result: ' . $result . PHP_EOL;
+	} else {
+		// Error handling here
+		// processing is not successful if thread dies
+		// when worked or working timeout exceeded
+		echo 'error' . PHP_EOL;
+	}
+}
+$thread->cleanup();